* 2.0: (172 commits)
  [Admin] Change name of official support section in menu and adjust translations
  Add all translations & change label
  Add business support CTA to admin panel
  [PromotionBundle] Add prefix to validation groups for promotion action/rule
  [PromotionBundle][PromotionRule] Add available type to validation message
  [PromotionBundle][PromotionAction] Add available type to validation message
  Update upgrade file
  [PromotionBundle] Split PromotionRuleValidator to PromotionRuleGroupValidator and PromotionRuleTypeValidator
  [PromotionBundle] Split PromotionActionValidator to PromotionActionGroupValidator and PromotionActionTypeValidator
  Add prefix sylius_ to validation groups
  Update upgrade file
  [PromotionBundle] Remove constraints from FormType to avoid double constraints
  [PromotionBundle] Move logic from PromotionRuleGroupValidator to PromotionRuleValidator
  [PromotionBundle] Move logic from PromotionActionGroupValidator to PromotionActionValidator
  [PromotionBundle] Add parameters validation groups for promotion action and promotion rule
  [ECS] Apply types order changes
  [ECS] Add union types ordering rule
  [ECS] Apply ecs fixes
  [ChannelsTest] use inArray to check locales
  [ApiBundle] Add contract tests for ProductTaxon resource
  ...

+++ src/Sylius/Bundle/PromotionBundle/Form/DataTransformer/PercentFloatToLocalizedStringTransformer.php
@@ -28,7 +28,7 @@ final class PercentFloatToLocalizedStringTransformer extends PercentToLocalizedS
      * @throws TransformationFailedException if the given value is not a string or
      *                                       if the value could not be transformed
      */
-    public function reverseTransform(mixed $value): int|float|null
+    public function reverseTransform(mixed $value): float|int|null
     {
         if (!is_numeric($value)) {
             return null;
